+++ src/app/api/tools/cowswap/route.ts
@@ -1,4 +1,4 @@
-import { parseQuoteRequest } from "@/src/app/api/tools/cowswap/util/protocol";
+import { parseQuoteRequest } from "@/src/app/api/tools/cowswap/util/parse";
 import { type NextRequest, NextResponse } from "next/server";
 import { orderRequestFlow } from "./orderFlow";
