+++ src/app/api/tools/cowswap/util/parse.ts
@@ -1,0 +1,83 @@
+import { getTokenDetails, TokenInfo } from "./tokens";
+import { extractAccountId } from "../../util";
+import { NextRequest } from "next/server";
+import {
+  OrderQuoteRequest,
+  OrderQuoteSideKindSell,
+  SigningScheme,
+} from "@cowprotocol/cow-sdk";
+import { getAddress, isAddress, parseUnits } from "viem";
+import { getSafeBalances } from "./balances";
+import { TokenBalance } from "../../safe-util";
+import { NATIVE_ASSET } from "./protocol";
+
+export interface ParsedQuoteRequest {
+  quoteRequest: OrderQuoteRequest;
+  chainId: number;
+}
+
+export async function parseQuoteRequest(
+  req: NextRequest,
+): Promise<ParsedQuoteRequest> {
+  // TODO - Add Type Guard on Request (to determine better if it needs processing below.)
+  const requestBody = await req.json();
+  console.log("Raw Request Body:", requestBody);
+  // TODO: Validate input with new validation tools:
+  const { sellToken, buyToken, chainId, sellAmountBeforeFee } = requestBody;
+  if (sellAmountBeforeFee === "0") {
+    throw new Error("Sell amount cannot be 0");
+  }
+
+  const { safeAddress: sender } = await extractAccountId(req);
+  
+  const [balances, buyTokenData] = await Promise.all([
+    getSafeBalances(chainId, sender),
+    getTokenDetails(chainId, buyToken),
+  ]);
+  const sellTokenData = sellTokenAvailable(balances, sellToken);
+
+  return {
+    chainId,
+    quoteRequest: {
+      sellToken: sellTokenData.address,
+      buyToken: buyTokenData.address,
+      sellAmountBeforeFee: parseUnits(
+        sellAmountBeforeFee,
+        sellTokenData.decimals,
+      ).toString(),
+      // TODO - change this when we want to enable buy orders.
+      kind: OrderQuoteSideKindSell.SELL,
+      // TODO - change this when we want to enable alternate recipients.
+      receiver: sender,
+      from: sender,
+      // manually add PRESIGN (since this is a safe);
+      signingScheme: SigningScheme.PRESIGN,
+    },
+  };
+}
+
+function sellTokenAvailable(
+  balances: TokenBalance[],
+  sellTokenSymbolOrAddress: string,
+): TokenInfo {
+  let balance: TokenBalance | undefined;
+  if (isAddress(sellTokenSymbolOrAddress)) {
+    balance = balances.find(
+      (b) =>
+        getAddress(b.tokenAddress || NATIVE_ASSET) === sellTokenSymbolOrAddress,
+    );
+  } else {
+    balance = balances.find(
+      (b) =>
+        b.token?.symbol.toLowerCase() ===
+        sellTokenSymbolOrAddress.toLowerCase(),
+    );
+  }
+  if (balance) {
+    return {
+      address: getAddress(balance.tokenAddress || NATIVE_ASSET),
+      decimals: balance.token?.decimals || 18,
+    };
+  }
+  throw new Error("Sell token not found in balances");
+}
